refactor TeamSignupSerializer
