+++ posthog/api/team.py
@@ -1,23 +1,17 @@
 from typing import Any, Dict
 
-import posthoganalytics
-from django.conf import settings
-from django.contrib.auth import login, password_validation
 from django.db import transaction
 from django.shortcuts import get_object_or_404
-from rest_framework import exceptions, generics, permissions, request, response, serializers, viewsets
+from rest_framework import exceptions, permissions, request, response, serializers, viewsets
 from rest_framework.decorators import action
-from rest_framework_extensions.routers import NestedRouterMixin
 
-from posthog.api.user import UserSerializer
-from posthog.models import Team, User
+from posthog.models import Team
 from posthog.models.utils import generate_random_token
 from posthog.permissions import (
     CREATE_METHODS,
     OrganizationAdminWritePermissions,
     OrganizationMemberPermissions,
     ProjectMembershipNecessaryPermissions,
-    UninitiatedOrCloudOnly,
 )
 
 
@@ -126,47 +120,3 @@ class TeamViewSet(viewsets.ModelViewSet):
         team.api_token = generate_random_token()
         team.save()
         return response.Response(TeamSerializer(team).data)
-
-
-class TeamSignupSerializer(serializers.Serializer):
-    first_name: serializers.Field = serializers.CharField(max_length=128)
-    email: serializers.Field = serializers.EmailField()
-    password: serializers.Field = serializers.CharField()
-    company_name: serializers.Field = serializers.CharField(max_length=128, required=False, allow_blank=True)
-    email_opt_in: serializers.Field = serializers.BooleanField(default=True)
-
-    def validate_password(self, value):
-        password_validation.validate_password(value)
-        return value
-
-    def create(self, validated_data, **kwargs):
-        is_first_user: bool = not User.objects.exists()
-        realm: str = "cloud" if getattr(settings, "MULTI_TENANCY", False) else "hosted"
-
-        company_name = validated_data.pop("company_name", validated_data["first_name"])
-        self._organization, self._team, self._user = User.objects.bootstrap(company_name=company_name, **validated_data)
-        user = self._user
-        login(
-            self.context["request"], user, backend="django.contrib.auth.backends.ModelBackend",
-        )
-
-        posthoganalytics.capture(
-            user.distinct_id,
-            "user signed up",
-            properties={"is_first_user": is_first_user, "is_organization_first_user": True},
-        )
-
-        posthoganalytics.identify(
-            user.distinct_id, properties={"email": user.email, "realm": realm, "ee_available": settings.EE_AVAILABLE},
-        )
-
-        return user
-
-    def to_representation(self, instance):
-        serializer = UserSerializer(instance=instance)
-        return serializer.data
-
-
-class TeamSignupViewset(generics.CreateAPIView):
-    serializer_class = TeamSignupSerializer
-    permission_classes = [UninitiatedOrCloudOnly]
